updates to properly adjust some dpdk settings in fatigue and a sleep in dpdk to prevent race condition that may have been causing a startup issue

+++ src/components/udp_source/socket/dpdk.hpp
@@ -45,13 +45,12 @@ private:
     struct rte_mempool* m_mbuf_pool;
     struct rte_eth_conf m_port_conf;
     uint16_t m_selected_port;
-    // struct udp::config m_config;
     std::string m_interface;
     std::string m_ip_addr;
-    size_t m_rx_ring_size{};
-    size_t m_num_mbufs{};
-    size_t m_mbuf_cache_size{};
-    size_t m_burst_size{};
+    uint32_t m_rx_ring_size{};
+    uint32_t m_num_mbufs{};
+    uint32_t m_mbuf_cache_size{};
+    uint32_t m_burst_size{};
     std::string m_socket_mem{};
 
 }; // class dpdk_udp
